requirements.txt
# Python依赖包
aiohttp>=3.8.0
aiofiles>=23.1.0
openai>=1.0.0
anthropic>=0.25.0
hepai>=1.0.0
PyYAML>=6.0
python-dateutil>=2.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
urllib3>=2.0.0

# Web界面依赖
Flask>=2.3.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
